+++ ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/ktor/server/QueryParameters.kt
@@ -12,13 +12,33 @@ suspend fun ApplicationCall.getParameterOrSendError(
     }
 }
 
+suspend fun ApplicationCall.getParametersOrSendError(
+    field: String
+) = parameters.getAll(field).also {
+    if (it == null) {
+        respond(HttpStatusCode.BadRequest, "Request must contains $field")
+    }
+}
+
 fun ApplicationCall.getQueryParameter(
     field: String
 ) = request.queryParameters[field]
 
+fun ApplicationCall.getQueryParameters(
+    field: String
+) = request.queryParameters.getAll(field)
+
 suspend fun ApplicationCall.getQueryParameterOrSendError(
     field: String
 ) = getQueryParameter(field).also {
+    if (it == null) {
+        respond(HttpStatusCode.BadRequest, "Request query parameters must contains $field")
+    }
+}
+
+suspend fun ApplicationCall.getQueryParametersOrSendError(
+    field: String
+) = getQueryParameters(field).also {
     if (it == null) {
         respond(HttpStatusCode.BadRequest, "Request query parameters must contains $field")
     }